moved applyBySpec to helpers and removed test method
